create response search class
