+++ packages/server/src/service/shorthands.ts
@@ -6,7 +6,6 @@ import { IAction, IStatement, ITerritory } from "@shared/types";
 import { IDbModel } from "@models/common";
 import { ModelNotValidError } from "@shared/types/errors";
 import { DbIndex, EntityClass } from "@shared/enums";
-import { regExpEscape } from "@common/functions";
 import Entity from "@models/entity/entity";
 import User from "@models/user/user";
 
@@ -40,77 +39,6 @@ export async function createEntity(
     throw new ModelNotValidError("");
   }
   return data.save(db.connection);
-}
-
-export async function filterEntitiesByWildcard(
-  db: Db,
-  entityClass: EntityClass | false,
-  entityClassExcluded: EntityClass[] | undefined,
-  entityLabel: string | false,
-  entityIds?: string[],
-  onlyTemplates?: boolean,
-  usedTemplate?: string
-): Promise<IEntity[]> {
-  let query = rethink.table(Entity.table);
-
-  if (entityIds && entityIds.length) {
-    query = query.getAll(rethink.args(entityIds)) as any;
-  }
-
-  if (entityClass) {
-    query = query.filter({
-      class: entityClass,
-    });
-  }
-
-  if (usedTemplate) {
-    query = query.filter({
-      usedTemplate: usedTemplate,
-    });
-  }
-
-  if (onlyTemplates) {
-    query = query.filter({
-      isTemplate: true,
-    });
-  }
-
-  if (entityClassExcluded) {
-    query = query.filter(function (row: RDatum) {
-      return rethink.and.apply(
-        rethink,
-        entityClassExcluded.map((c) => row("class").ne(c)) as [
-          RDatum<boolean>,
-          ...RDatum<boolean>[]
-        ]
-      );
-    });
-  }
-
-  if (entityLabel) {
-    let leftWildcard: string = "^",
-      rightWildcard: string = "$";
-
-    if (entityLabel[0] === "*") {
-      leftWildcard = "";
-      entityLabel = entityLabel.slice(1);
-    }
-
-    if (entityLabel[entityLabel.length - 1] === "*") {
-      rightWildcard = "";
-      entityLabel = entityLabel.slice(0, -1);
-    }
-
-    entityLabel = regExpEscape(entityLabel.toLowerCase());
-
-    query = query.filter(function (row: RDatum) {
-      return row("label")
-        .downcase()
-        .match(`${leftWildcard}${entityLabel}${rightWildcard}`);
-    });
-  }
-
-  return query.run(db.connection);
 }
 
 export async function deleteEntities(db: Db): Promise<WriteResult> {
